+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `StandaloneRangeDatePicker` component
+- Feature: Add `ListTable` component.
 - Feature: Add `DoubleArrowIcon` component.
 - Feature: Add `shadow`, `borderRadius`, `borderColor`, `borderWidth` props
   to `ArrowContainer` component.